subscribe_to_influx/requirements.txt
python-dotenv
quixstreams
influxdb-client